+++ nengo_extras/__init__.py
@@ -4,7 +4,8 @@ from .rc import rc
 # --- nengo_extras namespace (API)
 from .convnet import Conv2d, Pool2d
 from .neurons import FastLIF, SoftLIFRate
-from . import data, dists, graphviz, gui, networks, neurons, probe, vision
+from . import (
+    camera, data, dists, graphviz, gui, networks, neurons, probe, vision)
 
 __copyright__ = "2015-2018, Applied Brain Research"
 __license__ = "Free for non-commercial use; see LICENSE.rst"